AuthProvider useAuth refactor
