+++ src/components/Header/__tests__/AppMenu.test.tsx
@@ -2,19 +2,19 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { queryClient, render, screen } from 'test/test-utils';
 import { UseQueryResult } from '@tanstack/react-query';
 
-import * as UseAuthContext from 'providers/AuthProvider';
+import * as UseAuth from 'hooks/useAuth';
 import * as UseGetUser from 'api/useGetUser';
 import { userFixture1 } from '__fixtures__/users';
 
 import AppMenu from '../AppMenu';
 
 describe('AppMenu', () => {
-  const useAuthContextSpy = vi.spyOn(UseAuthContext, 'useAuthContext');
+  const useAuthSpy = vi.spyOn(UseAuth, 'useAuth');
   const useGetUserSpy = vi.spyOn(UseGetUser, 'useGetUser');
 
   beforeEach(() => {
     queryClient.clear();
-    useAuthContextSpy.mockReturnValue({
+    useAuthSpy.mockReturnValue({
       isAuthenticated: true,
     });
     useGetUserSpy.mockReturnValue({
@@ -62,7 +62,7 @@ describe('AppMenu', () => {
 
   it('should render unauthenticated content', async () => {
     // ARRANGE
-    useAuthContextSpy.mockReturnValue({
+    useAuthSpy.mockReturnValue({
       isAuthenticated: false,
     });
     useGetUserSpy.mockReturnValue({
